Provisional Reorg completed
