+++ r4/client/config.py
@@ -1,4 +1,5 @@
-from r4.client import S3
+from r4.client.s3 import S3
+
 regions = [
     S3.Region('us-east-1'),
     S3.Region('us-east-2'),
