+++ r4/client/rclient.py
@@ -5,16 +5,13 @@ import logging
 import threading
 
 from collections import deque
+from r4.client import AbstractFileManager, AbstractProvider
+from r4.client.s3 import S3
+from r4.client.r4 import R4
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 print = logger.info
-
-class AbstractFileManager(object):
-    def read(self, size):
-        raise NotYetImplemented()
-    def write(self):
-        raise NotYetImplemented()
 
 class UploadManagerFactory(object):
     '''
@@ -152,137 +149,6 @@ class DownloadManager(AbstractFileManager):
         with self.read_lock:
             logger.debug('blocked until download complete')
 
-class AbstractRegion(object):
-    def __init__(self, region_id):
-        if self.validate_region_id(region_id):
-            self.region_id = region_id
-        else:
-            self.region_id = None
-
-    def __str__(self):
-        return 'Region("%s")' % (self.region_id,)
-
-    def __repr__(self):
-        return 'Region("%s")' % (self.region_id,)
-
-    def validate_region_id(self, region_id):
-        return True
-
-class AbstractProvider(object):
-    def __str__(self):
-        raise NotYetImplemented()
-
-    def list(self):
-        raise NotYetImplemented()
-
-    def create(self, bucket_name, region):
-        raise NotYetImplemented()
-
-    def delete(self, bucket_name):
-        raise NotYetImplemented()
-
-    def delete_all(self):
-        raise NotYetImplemented()
-
-    def upload(self, bucket_name, file_key, file_obj):
-        raise NotYetImplemented()
-
-    def download(self, bucket_name, file_key, file_obj):
-        raise NotYetImplemented()
-
-class S3(AbstractProvider):
-    def __init__(self, region):
-        if isinstance(region, S3.Region):
-            self.region = region
-            self.s3 = boto3.resource('s3', region_name=region.region_id)
-            self.s3_client = boto3.client('s3', region_name=region.region_id)
-
-    class Region(AbstractRegion):
-        def validate_region_id(self, region_id):
-            return region_id in {
-                'ap-northeast-1',
-                'ap-northeast-2',
-                'ap-south-1',
-                'ap-southeast-1',
-                'ap-southeast-2',
-                'ca-central-1',
-                'eu-central-1',
-                'eu-east-1',
-                'eu-west-1',
-                'eu-west-2',
-                'sa-east-1',
-                'us-east-1',
-                'us-east-2',
-                'us-west-1',
-                'us-west-2',
-            }
-
-    def list(self):
-        for bucket in self.s3_client.list_buckets()['Buckets']:
-            yield bucket
-
-    def create(self, bucket_name):
-        try:
-            if self.region.region_id == 'us-east-1':
-                self.s3.Bucket(bucket_name).create()
-            else:
-                self.s3.Bucket(bucket_name).create(
-                    CreateBucketConfiguration = {
-                        'LocationConstraint': self.region.region_id,
-                    })
-            return True
-        except botocore.exceptions.ClientError as e:
-            if '%s' % (type(e),) == 'botocore.errorfactory.BucketAlreadyOwnedByYou':
-                print('bucket %s already exists' % (bucket_name,))
-            raise
-            return False
-
-
-    def delete(self, bucket_name):
-        # All objects (including all object versions and Delete Markers) in the 
-        #  bucket must be deleted before the bucket itself can be deleted.
-        # print('begin delete %s' % (bucket_name,))
-        self.s3.Bucket(bucket_name).objects.all().delete()
-        self.s3.Bucket(bucket_name).delete()
-        # print('end delete %s' % (bucket_name,))
-
-    def delete_all(self):
-        for bucket in self.list():
-            self.delete(bucket['Name'])
-
-    def upload(self, bucket_name, file_key, file_obj):
-        self.s3.Bucket(bucket_name).upload_fileobj(Fileobj=file_obj, Key=file_key)
-
-    def download(self, bucket_name, file_key, file_obj):
-        self.s3.Bucket(bucket_name).download_fileobj(Key=file_key, Fileobj=file_obj)
-
-class R4(AbstractProvider):
-    def __init__(self):
-        pass
-
-    class Region(AbstractRegion):
-        def validate_region_id(self, region_id):
-            # accept 'filesystem' and '<portnum>.LocalR4'
-            if region_id == 'filesystem':
-                return True
-            else:
-                if '.' in region_id:
-                    splits = region_id.split('.', 2)
-                    if len(splits) == 2:
-                        try:
-                            num = int(splits[0])
-                        except ValueError:
-                            return False
-                        return 1024 <= num <= 49151 and splits[1] == 'LocalR4'
-                    return False
-
-
-SUPPORTED_SERVICES = [
-    'Amazon Web Services S3',
-    'R4 Filesystem',
-    'R4 Local Server',
-    ]
-
 class Client(AbstractProvider):
     '''
     A Client is an AbstractProvider that accepts any and all regions and 
@@ -291,10 +157,6 @@ class Client(AbstractProvider):
     increasing performance through parallelization and increasing reliability by
     eliminating a single point of failure
     '''
-    default_regions = [
-        S3.Region('us-east-1'),
-        S3.Region('us-east-2'),
-    ]
     def __init__(self, regions):
         self.clients = {}
         if regions is None:
@@ -401,10 +263,10 @@ class Client(AbstractProvider):
         umf = UploadManagerFactory(data=data, fractional_upload=int(fractional_upload))
 
         for region in self.regions:
-            if isinstance(region, S3.Region):
-                client = 's3.'+region.region_id
-            elif isinstance(region, R4.Region):
+            if isinstance(region, R4.Region):
                 client = 'r4'
+            elif isinstance(region, S3.Region):
+                client = 's3.'+region.region_id
             else:
                 client = None
             if client is not None:
@@ -441,32 +303,3 @@ class Client(AbstractProvider):
 
         d.block_until_downloaded()
         return d.data
-
-
-if __name__ == '__main__':
-
-    s3 = Client(regions=[
-        S3.Region('us-east-1'),
-        S3.Region('us-east-2'),
-        S3.Region('us-west-1'),
-        S3.Region('us-west-2'),
-        ])
-    
-    bucket_name = 'io.r4.username03'
-
-    # print('Client.create("%s")' % (bucket_name,))
-    # s3.create(bucket_name)
-
-    # for bucket in s3.list():
-    #     print('Bucket? %s' % bucket['Name'])
-
-    upload_this = 'Hello AWS World. 4 Threads!'.encode('utf-8')
-    key = 'test_file'
-
-    s3.upload(bucket_name, key, upload_this, fractional_upload=1)
-    print('uploaded file %s to bucket %s' % (key, bucket_name))
-
-    data = s3.download(bucket_name, key, fractional_download=1)
-    print('downloaded file %s from bucket %s\n%s' % (key, bucket_name, data))
-
-    # s3.delete_all_buckets()
